Begin asio
